requirements.txt
fastapi==0.110.1
requests
uvicorn
PyJWT>=2.0.0